+++ common/src/ray.rs
@@ -7,7 +7,7 @@ pub struct Ray {
 }
 
 impl Ray {
-    pub fn at(self, t: f64) -> Point {
+    pub fn at(self, t: f32) -> Point {
         self.origin + self.direction * t
     }
 }
